Move the default providers configuration and classes in a dedicated directory
