+++ src/default_providers/Anthropic/settings-schema.json
@@ -1,0 +1,70 @@
+{
+  "$schema": "http://json-schema.org/draft-07/schema#",
+  "type": "object",
+  "properties": {
+    "temperature": {
+      "type": "number",
+      "description": "Amount of randomness injected into the response. Ranges from 0 to 1. Use temp closer to 0 for analytical / multiple choice, and temp closer to 1 for creative and generative tasks."
+    },
+    "topK": {
+      "type": "number",
+      "description": "Only sample from the top K options for each subsequent token. Used to remove \"long tail\" low probability responses. Defaults to -1, which disables it."
+    },
+    "topP": {
+      "type": "number",
+      "description": "Does nucleus sampling, in which we compute the cumulative distribution over all the options for each subsequent token in decreasing probability order and cut it off once it reaches a particular probability specified by top_p. Defaults to -1, which disables it. Note that you should either alter temperature or top_p, but not both."
+    },
+    "maxTokens": {
+      "type": "number",
+      "description": "A maximum number of tokens to generate before stopping."
+    },
+    "maxTokensToSample": {
+      "type": "number",
+      "description": "A maximum number of tokens to generate before stopping.",
+      "deprecated": "Use \"maxTokens\" instead."
+    },
+    "stopSequences": {
+      "type": "array",
+      "items": {
+        "type": "string"
+      },
+      "description": "A list of strings upon which to stop generating. You probably want `[\"\\n\\nHuman:\"]`, as that's the cue for the next turn in the dialog agent."
+    },
+    "streaming": {
+      "type": "boolean",
+      "description": "Whether to stream the results or not"
+    },
+    "anthropicApiKey": {
+      "type": "string",
+      "description": "Anthropic API key"
+    },
+    "apiKey": {
+      "type": "string",
+      "description": "Anthropic API key"
+    },
+    "anthropicApiUrl": {
+      "type": "string",
+      "description": "Anthropic API URL"
+    },
+    "modelName": {
+      "type": "string",
+      "deprecated": "Use \"model\" instead"
+    },
+    "model": {
+      "type": "string",
+      "description": "Model name to use"
+    },
+    "invocationKwargs": {
+      "type": "object",
+      "description": "Holds any additional parameters that are valid to pass to  {@link  * https://console.anthropic.com/docs/api/reference | }      * `anthropic.messages`} that are not explicitly specified on this class."
+    },
+    "streamUsage": {
+      "type": "boolean",
+      "description": "Whether or not to include token usage data in streamed chunks.",
+      "default": false
+    }
+  },
+  "additionalProperties": false,
+  "description": "Input to AnthropicChat class.",
+  "definitions": {}
+}
