+++ src/global.d.ts
@@ -7,3 +7,8 @@ declare module '*.svg' {
   const value: string;
   export default value;
 }
+
+declare module '*.md' {
+  const content: string;
+  export = content;
+}
